scaning isbn

+++ 301pro/app/src/main/java/com/example/a301pro/Functionality/ScanISBN.java
@@ -33,6 +33,7 @@ public class ScanISBN extends AppCompatActivity implements View.OnClickListener 
     private String Book_id;
     private String Name;
     private String person;
+    private String sta;
     protected FirebaseFirestore db;
     public static final String TAG = "SCAN_ISBN";
     /**
@@ -52,6 +53,7 @@ public class ScanISBN extends AppCompatActivity implements View.OnClickListener 
         ISBN = intent.getStringExtra("ISBN_CODE");
         Name = intent.getStringExtra("NAME");
         person = intent.getStringExtra("PERSON");
+        sta = intent.getStringExtra("STATUS");
 
         scanBtn = findViewById(R.id.scanBtn);
         scanBtn.setOnClickListener(this);
@@ -91,27 +93,9 @@ public class ScanISBN extends AppCompatActivity implements View.OnClickListener 
             if (result.getContents() != null) {
                 if (result.getContents().equals(ISBN)) {
 
-                    if(person.equals("Borrower")) {
-
-                        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-                        builder.setMessage("You have successfully borrowed the book");
-                        builder.setTitle("Scanning Result");
-                        builder.setNegativeButton("finish", new DialogInterface.OnClickListener() {
-                            @Override
-                            public void onClick(DialogInterface dialog, int which) {
-                                finish();
-                            }
-                        });
-
-                        AlertDialog dialog = builder.create();
-                        dialog.show();
+                    if(person.equals("Borrower") && sta.equals("Accepted")) {
 
                         final FirebaseFirestore db = FirebaseFirestore.getInstance();
-                        final CollectionReference collectionReference = db.collection("Users")
-                                .document(GetUserFromDB.getUserID())
-                                .collection("Borrowed");
-                        collectionReference.document(Book_id).update("sit", "Borrowed");
-
                         db.collection("userDict")
                                 .document(Name).get()
                                 .addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
@@ -123,21 +107,54 @@ public class ScanISBN extends AppCompatActivity implements View.OnClickListener 
                                                 .document(temp)
                                                 .collection("MyBooks")
                                                 .document(Book_id)
-                                                .update("status","Borrowed");
+                                                .get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
+                                            @Override
+                                            public void onSuccess(DocumentSnapshot documentSnapshot) {
+                                                String sss = documentSnapshot.getString("status");
+                                                if (sss.equals("Borrowed")){
+                                                    db.collection("Users")
+                                                            .document(GetUserFromDB.getUserID())
+                                                            .collection("Borrowed")
+                                                            .document(Book_id)
+                                                            .update("status", "Borrowed");
 
-                                        db.collection("Users")
-                                                .document(temp)
-                                                .collection("Request")
-                                                .document(Book_id)
-                                                .update("status", "Borrowed");
+                                                    AlertDialog.Builder builder = new AlertDialog.Builder(getApplicationContext());
+                                                    builder.setMessage("You have successfully borrowed the book");
+                                                    builder.setTitle("Scanning Result");
+                                                    builder.setNegativeButton("finish", new DialogInterface.OnClickListener() {
+                                                        @Override
+                                                        public void onClick(DialogInterface dialog, int which) {
+                                                            finish();
+                                                        }
+                                                    });
+
+                                                    AlertDialog dialog = builder.create();
+                                                    dialog.show();
+                                                }else{
+                                                    AlertDialog.Builder builder = new AlertDialog.Builder(getApplicationContext());
+                                                    builder.setMessage("The owner do not comfirm the lent");
+                                                    builder.setTitle("Scanning Result");
+                                                    builder.setNegativeButton("finish", new DialogInterface.OnClickListener() {
+                                                        @Override
+                                                        public void onClick(DialogInterface dialog, int which) {
+                                                            finish();
+                                                        }
+                                                    });
+
+                                                    AlertDialog dialog = builder.create();
+                                                    dialog.show();
+                                                }
+                                            }
+                                        });
                                     }
                                 });
 
                     }
-                    if (person.equals("Owner")){
+
+                    if (person.equals("Owner")&&sta.equals("Accepted")){
 
                         AlertDialog.Builder builder = new AlertDialog.Builder(this);
-                        builder.setMessage("You have successfully get the book back");
+                        builder.setMessage("You have successfully lent the book");
                         builder.setTitle("Scanning Result");
                         builder.setNegativeButton("finish", new DialogInterface.OnClickListener() {
                             @Override
@@ -152,34 +169,102 @@ public class ScanISBN extends AppCompatActivity implements View.OnClickListener 
                         final CollectionReference collectionReference = db.collection("Users")
                                 .document(GetUserFromDB.getUserID())
                                 .collection("Request");
-                        collectionReference.document(Book_id).delete();
+                        collectionReference.document(Book_id).update("status","Borrowed");
 
                         db.collection("Users")
                                 .document(GetUserFromDB.getUserID())
                                 .collection("MyBooks")
-                                .document(Book_id).update("status","Available");
+                                .document(Book_id).update("status","Borrowed");
 
-                        db.collection("Users")
+                    }
+
+                    if (person.equals("Borrower")&&sta.equals("Borrowed")){
+                        final FirebaseFirestore db = FirebaseFirestore.getInstance();
+                        final CollectionReference collectionReference = db.collection("Users")
                                 .document(GetUserFromDB.getUserID())
-                                .collection("MyBooks")
-                                .document(Book_id).update("borrower_name",null);
+                                .collection("Borrowed");
+                        collectionReference.document(Book_id).update("status","Available");
+                        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+                        builder.setMessage("You have successfully return the book");
+                        builder.setTitle("Scanning Result");
+                        builder.setNegativeButton("finish", new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface dialog, int which) {
+                                finish();
+                            }
+                        });
 
-                        db.collection("Library").document(Book_id).update("sit","Available");
+                        AlertDialog dialog = builder.create();
+                        dialog.show();
+                    }
+
+                    if (person.equals("Owner")&&sta.equals("Borrowed")){
 
                         db.collection("userDict")
                                 .document(Name).get()
                                 .addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
                                     @Override
                                     public void onSuccess(DocumentSnapshot documentSnapshot) {
-                                        String temp = documentSnapshot.getString("UID").toString();
+                                        final String temp = documentSnapshot.getString("UID").toString();
+
                                         db.collection("Users")
                                                 .document(temp)
                                                 .collection("Borrowed")
                                                 .document(Book_id)
-                                                .delete();
+                                                .get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
+                                            @Override
+                                            public void onSuccess(DocumentSnapshot documentSnapshot) {
+                                                String sss = documentSnapshot.getString("sit");
+                                                if (sss.equals("Available")){
+                                                    db.collection("Users")
+                                                            .document(temp)
+                                                            .collection("Borrowed")
+                                                            .document(Book_id)
+                                                            .delete();
+                                                    db.collection("Users")
+                                                            .document(GetUserFromDB.getUserID())
+                                                            .collection("Request")
+                                                            .document(Book_id)
+                                                            .delete();
+                                                    db.collection("Users")
+                                                            .document(GetUserFromDB.getUserID())
+                                                            .collection("MyBooks")
+                                                            .document(Book_id)
+                                                            .update("status","Available");
+                                                    db.collection("Library")
+                                                            .document(Book_id)
+                                                            .update("sit","Available");
+
+                                                    AlertDialog.Builder builder = new AlertDialog.Builder(getApplicationContext());
+                                                    builder.setMessage("You have successfully close the trade");
+                                                    builder.setTitle("Scanning Result");
+                                                    builder.setNegativeButton("finish", new DialogInterface.OnClickListener() {
+                                                        @Override
+                                                        public void onClick(DialogInterface dialog, int which) {
+                                                            finish();
+                                                        }
+                                                    });
+
+                                                    AlertDialog dialog = builder.create();
+                                                    dialog.show();
+                                                }else{
+                                                    AlertDialog.Builder builder = new AlertDialog.Builder(getApplicationContext());
+                                                    builder.setMessage("The Borrower do not return the book");
+                                                    builder.setTitle("Scanning Result");
+                                                    builder.setNegativeButton("finish", new DialogInterface.OnClickListener() {
+                                                        @Override
+                                                        public void onClick(DialogInterface dialog, int which) {
+                                                            finish();
+                                                        }
+                                                    });
+
+                                                    AlertDialog dialog = builder.create();
+                                                    dialog.show();
+                                                }
+                                            }
+                                        });
                                     }
                                 });
-
                     }
                 } else {
                     if (person.equals("Borrower")) {
